move the environment cleanup into the transitional State Tool
